Refactor: Optimize workspace modal
